Lock cloudwatch backend by SHA
This project has had some recent updates that make the usage pattern we
want feasible, but a new official version hasn't yet been released, and
the internal version in the `package.json` hasn't been updated either.

This led to some confusion since I installed the latest released version
& was following docs & reading code for a newer unreleased SHA.

The change here locks us to particular SHA on the master branch for now.
I've tested this by building the image & running it on a production
nexus instance on a on-standard port & sending messages to it with `nc`.
The metric appeared in cloudwatch as desired when I did this.